updated client to support docker 0.6.1 Fixed Issue #8

+++ src/main/java/com/kpelykh/docker/client/model/Image.java
@@ -22,10 +22,10 @@ public class Image {
     public long created;
 
     @JsonProperty("Size")
-    public int size;
+    public long size;
 
     @JsonProperty("VirtualSize")
-    public int virtualSize;
+    public long virtualSize;
 
     @Override
     public String toString() {
